[Chore #164697688] Refractored routes
