+++ src/middleware/middleware.js
@@ -1,0 +1,6 @@
+export default class Middleware {
+  static routeCallbacks(...methods) {
+    const callbacks = methods;
+    return callbacks;
+  }
+}
