+++ src/routes/accounts.js
@@ -1,28 +1,13 @@
 import accountsController from '../controllers/accounts';
 import router from './router';
-import { ValidateAccountInput, AuthenticateAccount } from '../middleware/accounts';
-import { AuthenticateUsers } from '../middleware/users';
+import accountsMiddleware from '../middleware/accounts';
 
-router.post('/accounts', (...args) => {
-  ValidateAccountInput.createAccount(...args);
-}, (...args) => {
-  AuthenticateUsers.clients(...args);
-}, accountsController.createAccount);
+router.post('/accounts', accountsMiddleware.createAccount(), accountsController.createAccount);
 
-router.delete('/accounts/:account_number', (...args) => {
-  ValidateAccountInput.deleteAccount(...args);
-}, (...args) => {
-  AuthenticateUsers.admin(...args);
-}, (...args) => {
-  AuthenticateAccount.account(...args);
-}, accountsController.deleteAccount);
+router.delete('/accounts/:account_number', accountsMiddleware.deleteAccount(),
+  accountsController.deleteAccount);
 
-router.patch('/account/:account_number', (...args) => {
-  ValidateAccountInput.updateAccountStatus(...args);
-}, (...args) => {
-  AuthenticateUsers.admin(...args);
-}, (...args) => {
-  AuthenticateAccount.account(...args);
-}, accountsController.updateStatus);
+router.patch('/account/:account_number', accountsMiddleware.updateAccountStatus(),
+  accountsController.updateStatus);
 
 export default router;
