+++ src/routes/users.js
@@ -1,41 +1,17 @@
 import userController from '../controllers/users';
 import router from './router';
-import { ValidateUserInput, AuthenticateUsers } from '../middleware/users';
+import userMiddleware from '../middleware/users';
 
-router.post('/auth/signup', (...args) => {
-  ValidateUserInput.signUp(...args);
-}, (...args) => {
-  AuthenticateUsers.signUp(...args);
-}, userController.signUp);
+router.post('/auth/signup', userMiddleware.signupClients(), userController.signUp);
 
-router.post('/auth/signin', (...args) => {
-  ValidateUserInput.signIn(...args);
-}, (...args) => {
-  AuthenticateUsers.signIn(...args);
-}, userController.signIn);
+router.post('/auth/signin', userMiddleware.signinClients(), userController.signIn);
 
-router.post('/auth/signup/admin', (...args) => {
-  ValidateUserInput.adminStaff(...args);
-}, (...args) => {
-  AuthenticateUsers.signUpAdmin(...args);
-}, userController.signUpAdmin);
+router.post('/auth/signup/admin', userMiddleware.signupAdmin(), userController.signUpAdmin);
 
-router.post('/auth/signin/admin', (...args) => {
-  ValidateUserInput.adminStaff(...args);
-}, (...args) => {
-  AuthenticateUsers.signInAdmin(...args);
-}, userController.signinAdminStaff);
+router.post('/auth/signin/admin', userMiddleware.signinAdmin(), userController.signinAdminStaff);
 
-router.post('/auth/signin/staff', (...args) => {
-  ValidateUserInput.adminStaff(...args);
-}, (...args) => {
-  AuthenticateUsers.authSignInStaff(...args);
-}, userController.signinAdminStaff);
+router.post('/auth/signin/staff', userMiddleware.signinStaff(), userController.signinAdminStaff);
 
-router.post('/auth/signup/staff', (...args) => {
-  ValidateUserInput.adminStaff(...args);
-}, (...args) => {
-  AuthenticateUsers.signUpStaff(...args);
-}, userController.signUpStaff);
+router.post('/auth/signup/staff', userMiddleware.signupStaff(), userController.signUpStaff);
 
 export default router;
